Bugfix: Allow for more than 65535 minutes of inactivity.

+++ src/account.cpp
@@ -40,7 +40,7 @@ const void Account::set_minutes(uint16_t minutes)
     _minutes = minutes;
 }
 
-const uint16_t Account::get_minutes() const
+const uint64_t Account::get_minutes() const
 {
     return _minutes;
 }
